refactor(MainPage): MainPage 컴포넌트를 Root page 컴포넌트로 이동
